fix(tests): fix assertions on default beverage
This fixes an issue where integration tests that would otherwise
succeed, locally, fail close to the end because the user executing
the tests has set the DEIS_DRINK_OF_CHOICE environment variable,
which interferes with the assertion that the default beverage to
be enjoyed whilst scaling is coffee.  This corrects that problem
by unsetting the DEIS_DRINK_OF_CHOICE environment variable prior
to testing.