feat(controller): add health check feature to app config
This feature extends the publisher and the controller to check if the
user specified a health check for the application. If they did,
publisher will only publish the app if the application passes the health
check, and the controller will roll back if a deployment fails to pass
this health check.

By default, the expected status code is 200 unless the user specifies
otherwise. This variable is non-configurable on the platform side as
both the controller and the publisher have to rely on this value, and I
could not find a happy medium for setting up a shared config variable.