fix(builder): only build slugbuilder/slugrunner when necessary
On start, the slugbuilder and slugrunner images are built from
the builder's source directory. This is unnecessary on restarts,
and prevents optimizations where the slugbuilder and slugrunner images
are already pre-cached (for example, by bind-mounting a host volume
into builder with the images).

This commit simply wraps these `docker build` instructions in
`docker history` checks to see if the images exist before
attempting to build them.