fix(router): add set-misc module to nginx build
The commit c7c895c fixes the affinityArg hashing issue by using the
`set_random` directive on the nginx configuration, but forgets to
include the module in which it is declared. This commit fixes this,
while also tries to make the build script a little cleaner and
include integrity checks for all downloads.