fix(logger): use one ticker for all time-related operations
In a separate goroutine, we were sleeping for the exact same time as
publishInterval was set to. As @kalbasit mentioned in [a comment][1],
it's better to use a ticker anyways. This removes one extra underlying
system call to clock_gettime() by making both publishKeys() and the
log drain poll share the same ticker event.

[1]: https://github.com/deis/deis/pull/3795#discussion_r31878237