fix(controller): convert config to correct types
config.cpu will always be a integer

config.memory, config.values, config.tags will always be
a string

Previously the controller would take and save any type,
provided it passed validation. This was problematic because
it could lead to mixed or unexpected types.

Examples:

This config would be valid:

```
{
  "values": {
    "TEST1": true,
    "TEST2": 123,
    "TEST3": "test"
  }
}
```

You could get a response like this:

```
{
  "cpu": {
    "web": 1024,
    "worker": "512"
  }
}
```