feat(deisctl): extend SSH to allow exec and docker
This feature extends 'deis ssh' to allow for an arbitrary command.
If specified, the command will be run as an SSH Exec call, instead
of opening a new shell.

Further, this adds 'deis dock', which is a convenience wrapper
for 'deis ssh THING docker exec -it THING COMMAND'. This now
becomes 'deis dock THING COMMAND'.