chore(deisctl): refactor config/etcd code
In order to reuse the code for accessing a
back-end config store (e.g. etcd) in the
new upgrade feature, and in order to test
that same code, the previous commit made
some compromises that might have affected
the long-term maintainability of deisctl.
This is a follow-up commit to refactor
some of that.  It establishes cleaner
boundaries between config back-end
model, interface, and that interface's
two implementations-- the existing one
based on etcd and a new, in-memory "mock"
used for testing.