fix(create_bucket): check for existence of None
There are small implications between `if not x` and `if x is None` in
Python:

 - `if not x` checks if the value is false (empty list, `None`, `False`)
 - `if x is None` checks if x is equal to the literal value `None`.

conn.lookup returns `None` if no bucket exists. This code is explicitly
stated in their docs as the proper way to check if a bucket does not
exist.

See: http://boto.readthedocs.org/en/latest/s3_tut.html