feat(contrib/linode): standardize scripts and provide IP discovery options
Standardized and normalized the python functions used during CoreOS setup
(e.g. Linode api calls and logging). Also, I added multiple methods of
providing IPs when applying a firewall to the cluster. Namely, IPs can now
be provided via a display-group, manually, or using the discovery URL.
This is useful particularly when running this command outside of the
cluster, but also when adding nodes to an existing cluster (because
the new nodes will be added as proxies).