fix(controller): use django HttpResponse for logs
Django Rest Framework will wrap the string with quotes because
of the JSONRenderer. Django's HttpResponse class handles the
application logs as we expect it to.

This scrubs the "No logs for appname" message we would display when
there are no logs to show, but in reality this is the correct way to
return a 204, and that message was never displayed to users because
we inject "user created initial release" when a user first runs
`deis apps:create`. Django scrubs the response content if the
status code is a 204 to comply with RFC 2616, section 4.3.

See https://github.com/django/django/blob/53ccffdb8c8e47a4d4304df453d8c79a9be295ab/django/http/utils.py#L13-L16