ref(client): improve client error message
This PR improves the resultant client error message they get when the returned response is
incorrect. This is especially important for determining the error message

Old message when using a v1 client with a v2 cluster:

```
http://deis.10.245.1.3.xip.io does not appear to be a valid Deis controller. Make sure that the
Controller URI is correct and the server is running.
```

New message:

```
http://deis.10.245.1.3.xip.io/v1/ does not appear to be a valid Deis controller. Make sure that the
Controller URI is correct and the server is running.
```

Notice the appended /v1/ path, which users can assume this is a v1 client.