feat(contrib): Add script to restart all apps.
This is usefull after a reinstall of a cluster with stateless storage.
When the new cluster is started it will still remember the requested
scale of all applications, but all containers will be in "destroyed"
state.
This means that a `deis ps:restart` does not work and so manual
rescaling needs to be done.
This is a tedious and error-prone task if you have many applications,
so this little script takes care of that.